Add `contents` option
The displayed contents for each file are fetched from the specified `url`
option.

It is also convenient to have the ability to specify the raw file contents
directly. This is espcially true for smaller files, or if it is difficult to
host your file snippets under some remote URL.

This commit adds a `contents` option to which can be used to specify the
file contents inline.

\# Usage

  ```
  <script>
    const files = [
      {
        path: 'lib/file.js',
        contents: 'const foo = () => {\n  alert('foo');\n}',
        language: 'javascript'
      }
    ];

    new VanillaTreeViewer('some-id', files).render();
  </script>
  ```

\# Behavior

* This commit adds a validation to ensure at least one of `url` or `contents`
  is specified.
* This option takes precedence over `url` if both are specified