Use `.vtv-root` instead of `.vtv` for the top level CSS class name
Up till now, we've mounted the `VanillaTreeViewer` class on top of the user
generated mount node and given it a top-level class of `.vtv`

    <div id="some-user-mount-node">
      <div class="vtv">
        ...
      </div>
    </div>

As part of an upcoming change we plan on having the user generated mount use
a common, consistent CSS class to identify the mount nodes. It would be
much more convenient to have the user be able to use `.vtv`, from a simplicity
and usability perspective.

To accommodate this, we're pre-emptively updating the root `<div>` to use
something other than `.vtv`: `.vtv-root`

This is not a "Breaking Change" because we advise users against using the
internal CSS classes for styling or other functions. But users may be doing
this anyway 👀