[Breaking Change] Parse structure and options from the DOM directly
This change represents a significant update to the way we initialize
and configure `VanillaTreeViewer`. It builds off of over a year's worth
off feedback and learning from `VanillaTreeViewer` `v1`.

Historically we've passed in a set of options to the constructor:

    new VanillaTreeViewer(id, files)

where `files` is the configuration object listing all files.

This approach is clunky for most users as it requires having to manually
build out a configuration object and pass it in via JavaScript.

The updated approach reads all configuration from the DOM directly:

    <ol class='vtv' data-language="javascript">

      <!-- File 1: src/index.js -->
      <li data-path="src/index.js" >
    import Foo from './foo';
    export { Foo.bar }
      </li>

      <!-- File 2: package.json -->
      <!-- File contents will be fetched from the specified `url` -->
      <!-- Any file-level options here will override parent-level options (e.g. `data-language`)  -->
      <li
        data-path="package.json"
        data-url="https://raw.githubusercontent.com/axios/axios/master/package.json"
        data-language="json">
      </li>
    </ol>

This has a few advantages:

  1. It's more intuitive for users - the schema and structure of the
     file list matches how people tend to think about a list of files.

  2. It allows users to configure `VanillaTreeViewer` as part of their
     HTML templates and rendering. Essentially it allows them to work
     in HTML directly instead of JavaScript.

  3. The user can now make a single call to instantiate all instances
     at once: `VanillaTreeViewer.renderAll();`. It also no longer requires
     initializing a `VanillaTreeViewer` instance.

This is a breaking change since it heavily changes the public API. The scope
of this change is:

  * Parsing the DOM for configuration
  * Updating / adding tests
  * Updating `README` documentation
  * Updating the Demo page