Align `Header` action buttons
A recent commit (84d8ac8) added the "Wrap Text" toggle button.

The `Header` now contains multiple actions, and each was using its own
individual CSS styling. This further cause the issue of visual
mis-alignment at some widths.

This change cleans up the CSS by moving all styling for actions into a
common selector (`vtv__code-panel-action`) and active state
(`vtv__code-panel-action--active`).

The `Logo` was also restructured to be an `<a>` wrapped inside a
`<button>` instead of the previous approach of a `<button>` wrapped in an
`<a>`. This keeps all actions consisten in their structure.